Feat/cart to Develop

+++ src/main/java/com/sparta/palpaleats/domain/user/dto/UserOrderResponseDto.java
@@ -17,7 +17,7 @@ public class UserOrderResponseDto {
 
     private Long orderId;
 
-    private Integer totalPrice;
+    private Long totalPrice;
 
     private String paymentMethod;
